Add files via upload
Created visualizations and log transformations for the target variable.  Moved summary to the top of the document.  Team needs to review their sections and fix any bugs.  Team needs to provide explanations to their code.  There are large blocks of code that do not have any explanation as to what it is for.